feat!: Rename configuration type
BREAKING CHANGE: Renamed `SetupConfiguration` to `Configuration` to indicate its no longer just
during a 'setup' phase
