+++ src/features/search/category-search/category-search.spec.ts
@@ -1,11 +1,11 @@
 import { faker } from '@faker-js/faker';
 import { HttpResponse, http } from 'msw';
 import { describe, expect, test } from 'vitest';
-import { createSetupConfigMock } from '../../../initialize/configuration.mock';
 import { mockRequest } from '../../../shared/mock-request.mock';
 import { createSearchResponseMock } from '../search-response.mock';
 import { categorySearch } from './category-search';
 import { createCategorySearchOptionsMock } from './category-search-options.mock';
+import { createSetupConfigMock } from '../../../shared/configuration.mock';
 
 describe('Category Search API', () => {
   const config = createSetupConfigMock();
