+++ src/features/search/content-search/content-search.spec.ts
@@ -1,10 +1,10 @@
 import { HttpResponse, http } from 'msw';
 import { describe, expect, test } from 'vitest';
-import { createSetupConfigMock } from '../../../initialize/configuration.mock';
 import { mockRequest } from '../../../shared/mock-request.mock';
 import { createSearchResponseMock } from '../search-response.mock';
 import { contentSearch } from './content-search';
 import { createContentSearchOptionsMock } from './content-search-options.mock';
+import { createSetupConfigMock } from '../../../shared/configuration.mock';
 
 describe('Content Search API', () => {
   const config = createSetupConfigMock();
