+++ src/features/search/content-search/content-search.ts
@@ -1,4 +1,4 @@
-import type { Configuration } from '../../../initialize/configuration.type';
+import type { Configuration } from '../../../shared/configuration.type';
 import { SEARCH_ENDPOINT_PROD } from '../../../shared/constants';
 import { buildApiUrl } from '../../../shared/url-builders';
 import type { ContentSearchRequestParameters } from '../search-request.type';
