--- src/initialize/initialize.ts
@@ -1,6 +1,0 @@
-import { setConfig } from '../shared/config';
-import type { Configuration } from './configuration.type';
-
-export function initialize(config: Configuration) {
-  setConfig(config);
-}
