+++ src/shared/configuration.mock.ts
