+++ src/shared/configuration.type.ts
